Rename youtube to playlist in react and redux files
