+++ App/Sagas/YoutubeSaga.js
@@ -1,6 +1,6 @@
 import { put, call } from 'redux-saga/effects'
 
-import YoutubeActions from 'App/Stores/Youtube/Actions'
+import PlaylistActions from 'App/Stores/Playlist/Actions'
 import youtubeService from 'App/Services/YoutubeService'
 import apiErrorTransform from 'App/utils/apiErrorTransform'
 
@@ -8,8 +8,8 @@ export function* fetchPlaylistsSaga(action) {
   const result = yield call(youtubeService.getPlaylists, { pageToken: action.pageToken })
 
   if (result.ok) {
-    yield put(YoutubeActions.fetchPlaylistsSuccess(result.data, action.fetchPageNumber))
+    yield put(PlaylistActions.fetchPlaylistsSuccess(result.data, action.fetchPageNumber))
   } else {
-    yield put(YoutubeActions.fetchPlaylistsFailure(apiErrorTransform(result)))
+    yield put(PlaylistActions.fetchPlaylistsFailure(apiErrorTransform(result)))
   }
 }
